add env resolver to client.ts too
